requirements.txt
faunadb==4.1.0
python_telegram_bot==13.6
python-dotenv==0.18.0
telegram==0.0.1
